% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcR.R
\name{ipcr_data}
\alias{ipcr_data}
\title{ipcr_data}
\usage{
ipcr_data()
}
\value{
Complete dataset
}
\description{
This function allows you to find and display the dataset on Jaccard similarity coefficients for innovation dynamics in the pharmaceutical industry in China. .
}
\examples{
myData <- ipcr_data()
}
