% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcR.R
\name{ipcr_visual}
\alias{ipcr_visual}
\title{ipcr_visual}
\usage{
ipcr_visual(chart = "line_1", class = "b14s", title = TRUE)
}
\arguments{
\item{chart}{Type of charts}

\item{class}{4-digit class of patent}

\item{title}{Chart title, set by default to TRUE}
}
\value{
Chosen Graph
}
\description{
This function allows you to create 4 sorts of visuals: line, bar, box and point charts.
}
\examples{
ipcr_visual()
}
